Shrink NodeCoords to 16 B/record and permute it in place
Two changes to the node coordinate store, both aimed at Pass B's peak RSS
(measured as the largest identified structure on a country-sized extract).

1. In-place permutation. `NodeCoords::build` used to allocate a second
   Vec and scatter each record into `data[mphf.hash(id)]`, holding both
   arrays live for the whole loop — a Vec's buffer isn't freed until its
   IntoIter drops, so moving elements out one at a time reclaims nothing
   along the way. Since a minimal perfect hash is a bijection onto
   0..len, the target layout is just `records` permuted, so it's applied
   by cycle-following with a len-bit visited bitset (~9 MB) instead.

2. 24 B -> 16 B records. `(i64, f32, f32, bool)` padded to 24 bytes — 8
   of them spent carrying one bit. `shared` moves to a sidecar bitset,
   and coordinates move from f32 degrees to i32 decimicrodegrees, taken
   straight from osmpbf's decimicro_lon/decimicro_lat. That's the PBF's
   own native integer form, so it removes a lossy f64->f32 narrowing
   rather than adding a conversion: stored coordinates are now exact
   (verified bit-identical against osmium ground truth) where f32
   previously carried up to ~2 m of error at the longitude range's top
   end. The builder accumulates the same 16 B shape so `finish` hands
   its buffers straight to `build` with no shrink-copy.

Note this changes emitted geometry values: they get more precise, so a
diff against an older import will show sub-meter differences.

Measured on germany-latest.osm.pbf + configs/tilda (75.85M referenced
nodes), --threads 8: wall clock 6:16.53 -> 5:21.47 (-14.6%), peak RSS
10.64 GB -> 9.15 GB (-14.0%). Row counts identical across all four
tables; materialize phase 38.5s -> 34.0s.

+++ src/geom/materialize.rs
@@ -182,7 +182,7 @@ where
     let (node_ids, node_rows) = if plan.any_way_graph {
         let endpoints: FxHashSet<i64> = ctx.way_refs.endpoints();
         let (ids, rows) = assign_node_ids(&ctx.node_coords, &endpoints, &ctx.selected);
-        let node_rows = rows.into_iter().map(|(id, osm_id, lon, lat)| build_node_row(id, osm_id, lon as f64, lat as f64)).collect();
+        let node_rows = rows.into_iter().map(|(id, osm_id, lon, lat)| build_node_row(id, osm_id, lon, lat)).collect();
         (ids, node_rows)
     } else {
         (FxHashMap::default(), Vec::new())
